AG-9402 - Move api explorer components and utils into folders
