+++ packages/ag-charts-website/src/features/api-explorer/components/ChartsApiExplorer.module.scss
@@ -1,4 +1,4 @@
-@use '../../design-system/' as *;
+@use '../../../design-system/' as *;
 
 .container {
     height: calc(100vh - 210px);
