+++ packages/ag-charts-website/src/features/api-explorer/components/CodeView.module.scss
@@ -1,4 +1,4 @@
-@use '../../design-system/' as *;
+@use '../../../design-system/' as *;
 
 .code {
     display: flex;
