+++ packages/ag-charts-website/src/features/api-explorer/components/Editors.module.scss
@@ -1,4 +1,4 @@
-@use '../../design-system/' as *;
+@use '../../../design-system/' as *;
 
 .stringEditor {
     width: 100%;
