+++ packages/ag-charts-website/src/features/api-explorer/components/Options.module.scss
@@ -1,5 +1,5 @@
 @use 'sass:map';
-@use '../../design-system/' as *;
+@use '../../../design-system/' as *;
 
 .options {
     position: absolute;
