+++ packages/ag-charts-website/src/pages/explorer.astro
@@ -2,7 +2,7 @@
 import type { Framework } from '@ag-grid-types';
 import APIViewLayout from '@layouts/APIViewLayout.astro';
 import { getJsonFromDevFile } from '@utils/getJsonFromDevFile';
-import { ChartsApiExplorer } from '@features/api-explorer/ChartsApiExplorer';
+import { ChartsApiExplorer } from '@features/api-explorer/components/ChartsApiExplorer';
 import { pathJoin } from '@utils/pathJoin';
 import { SITE_BASE_URL } from '@constants';
